+++ pom.xml
@@ -9,9 +9,9 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.example</groupId>
-	<artifactId>Registration_Login</artifactId>
+	<artifactId>ImageIOT</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>Registration_Login</name>
+	<name>ImageIOT</name>
 	<description>Demo project for Spring Boot</description>
 	<url/>
 	<licenses>
@@ -102,6 +102,10 @@
     		<groupId>com.sendgrid</groupId>
     		<artifactId>sendgrid-java</artifactId>
     		<version>4.6.8</version>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-validation</artifactId>
 		</dependency>
 	</dependencies>
 	<build>
