+++ src/main/java/com/example/ImageIOT/RegistrationLoginApplication.java
@@ -1,0 +1,15 @@
+package com.example.ImageIOT;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@SpringBootApplication
+@EnableJpaAuditing
+public class RegistrationLoginApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(RegistrationLoginApplication.class, args);
+	}
+
+}
