+++ src/main/java/com/example/ImageIOT/dto/LoginRequestDTO.java
@@ -1,0 +1,16 @@
+package com.example.ImageIOT.dto;
+
+import lombok.Data;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+
+@Data
+public class LoginRequestDTO {
+    
+    @NotBlank
+    @Email
+    private String email;
+
+    @NotBlank
+    private String password;
+}
