+++ src/main/java/com/example/ImageIOT/dto/UserResponseDTO.java
@@ -1,0 +1,19 @@
+package com.example.ImageIOT.dto;
+
+import com.example.ImageIOT.enums.AuthProvider;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class UserResponseDTO {
+
+    
+    private String name;
+    private String email;
+    private Long phonenumber;
+    private AuthProvider authProvider;
+}
