+++ src/main/java/com/example/ImageIOT/enums/AuthProvider.java
@@ -1,0 +1,9 @@
+package com.example.ImageIOT.enums;
+
+public enum AuthProvider {
+
+    LOCAL, 
+    GOOGLE, 
+    FACEBOOK, 
+    GITHUB
+}
