+++ src/main/java/com/example/ImageIOT/enums/Role.java
@@ -1,0 +1,7 @@
+package com.example.ImageIOT.enums;
+
+public enum Role {
+    
+    USER, 
+    ADMIN
+}
