+++ src/main/java/com/example/ImageIOT/exception/CustomException.java
@@ -1,0 +1,19 @@
+package com.example.ImageIOT.exception;
+
+public class CustomException extends RuntimeException {
+    private final int errorCode;
+
+    public CustomException(String message) {
+        super(message);
+        this.errorCode = 0;
+    }
+
+    public CustomException(String message, int errorCode) {
+        super(message);
+        this.errorCode = errorCode;
+    }
+
+    public int getErrorCode() {
+        return errorCode;
+    }
+}
