+++ src/main/java/com/example/ImageIOT/filter/JwtFilter.java
@@ -1,0 +1,64 @@
+package com.example.ImageIOT.filter;
+
+import com.example.ImageIOT.service.CustomUserDetailsService;
+import com.example.ImageIOT.util.JwtUtil;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.OncePerRequestFilter;
+import java.io.IOException;
+
+@Component
+@Slf4j
+public class JwtFilter extends OncePerRequestFilter {
+
+    @Autowired
+    private JwtUtil jwtUtil;
+
+    @Autowired
+    private CustomUserDetailsService customUserDetailsService;
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain chain)
+            throws ServletException, IOException {
+        String authHeader = request.getHeader("Authorization");
+        if (authHeader == null || !authHeader.startsWith("Bearer ")) {
+            chain.doFilter(request, response);
+            return;
+        }
+
+        String jwt = authHeader.substring(7);
+        try {
+            String username = jwtUtil.extractUsername(jwt);
+            if (username == null || SecurityContextHolder.getContext().getAuthentication() != null) {
+                chain.doFilter(request, response);
+                return;
+            }
+
+            UserDetails userDetails = customUserDetailsService.loadUserByUsername(username);
+            if (!jwtUtil.validateToken(jwt, userDetails.getUsername())) {
+                log.warn("Invalid JWT token");
+                chain.doFilter(request, response);
+                return;
+            }
+
+            var authorities = jwtUtil.extractRoles(jwt);
+            UsernamePasswordAuthenticationToken authToken = new UsernamePasswordAuthenticationToken(
+                    userDetails, null, authorities);
+            authToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
+            SecurityContextHolder.getContext().setAuthentication(authToken);
+        } catch (Exception e) {
+            log.error("JWT extraction/validation failed: {}", e.getMessage());
+        }
+        chain.doFilter(request, response);
+    }
+
+}
