+++ src/main/java/com/example/ImageIOT/model/User.java
@@ -1,0 +1,46 @@
+package com.example.ImageIOT.model;
+
+
+
+import java.util.Set;
+
+import com.example.ImageIOT.enums.AuthProvider;
+import com.example.ImageIOT.enums.Role;
+
+import jakarta.annotation.Generated;
+import jakarta.persistence.ElementCollection;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+@Table(name = "users")
+public class User {
+    
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String name;
+    private String email;
+    private Long phonenumber;
+    private String password;
+
+    @Enumerated(EnumType.STRING)
+    @ElementCollection(fetch = FetchType.EAGER)
+    private Set<Role> roles;
+
+    @Enumerated(EnumType.STRING)
+    private AuthProvider authProvider;
+
+}
