+++ src/main/java/com/example/ImageIOT/security/SendGridConfig.java
@@ -1,0 +1,18 @@
+package com.example.ImageIOT.security;
+
+import com.sendgrid.SendGrid;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SendGridConfig {
+
+    @Value("${sendgrid.api-key}")
+    private String apiKey;
+
+    @Bean
+    public SendGrid sendGrid() {
+        return new SendGrid(apiKey);
+    }
+}
