+++ src/main/java/com/example/ImageIOT/service/CustomUserDetailsService.java
@@ -1,0 +1,34 @@
+package com.example.ImageIOT.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+import com.example.ImageIOT.exception.CustomException;
+import com.example.ImageIOT.model.User;
+import com.example.ImageIOT.repository.UserRepository;
+
+import java.util.Collections;
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class CustomUserDetailsService implements UserDetailsService {
+
+    private final UserRepository userRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        List<User> users = userRepository.findByEmail(email);
+        if (users.isEmpty()) {
+           throw new CustomException("User not found with email: " + email, 404);
+        }
+        User user = users.get(0);
+        return new org.springframework.security.core.userdetails.User(
+                user.getEmail(),
+                user.getPassword(),
+                Collections.emptyList());
+    }
+}
