+++ src/main/java/com/example/ImageIOT/service/RefreshTokenService.java
@@ -1,0 +1,50 @@
+package com.example.ImageIOT.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+
+import com.example.ImageIOT.exception.CustomException;
+import com.example.ImageIOT.model.RefreshToken;
+import com.example.ImageIOT.model.User;
+import com.example.ImageIOT.repository.RefreshTokenRepository;
+
+import java.time.Instant;
+import java.util.Optional;
+import java.util.UUID;
+
+@Service
+@RequiredArgsConstructor
+public class RefreshTokenService {
+
+    @Value("${jwt.refresh.expiration}")
+    private Long refreshTokenDurationMs;
+
+    private final RefreshTokenRepository refreshTokenRepository;
+
+    public Optional<RefreshToken> findByToken(String token) {
+        return refreshTokenRepository.findByToken(token);
+    }
+
+    public RefreshToken createRefreshToken(User user) {
+        RefreshToken refreshToken = new RefreshToken();
+        refreshToken.setUser(user);
+        refreshToken.setExpiryDate(Instant.now().plusMillis(refreshTokenDurationMs));
+        refreshToken.setToken(UUID.randomUUID().toString());
+
+        refreshToken = refreshTokenRepository.save(refreshToken);
+        return refreshToken;
+    }
+
+    public RefreshToken verifyExpiration(RefreshToken token) {
+        if (token.getExpiryDate().compareTo(Instant.now()) < 0) {
+            refreshTokenRepository.delete(token);
+            throw new CustomException("Refresh token was expired. Please make a new signin request", 401);
+        }
+        return token;
+    }
+
+    public void deleteByUser(User user) {
+        refreshTokenRepository.deleteByUser(user);
+    }
+}
