--- src/main/java/com/example/Registration_Login/RegistrationLoginApplication.java
@@ -1,15 +1,0 @@
-package com.example.Registration_Login;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@SpringBootApplication
-@EnableJpaAuditing
-public class RegistrationLoginApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(RegistrationLoginApplication.class, args);
-	}
-
-}
