--- src/main/java/com/example/Registration_Login/dto/LoginRequestDTO.java
@@ -1,10 +1,0 @@
-package com.example.Registration_Login.dto;
-
-import lombok.Data;
-
-@Data
-public class LoginRequestDTO {
-    
-    private String email;
-    private String password;
-}
