--- src/main/java/com/example/Registration_Login/dto/OtpRequestDTO.java
@@ -1,9 +1,0 @@
-package com.example.Registration_Login.dto;
-
-import lombok.Data;
-
-@Data
-public class OtpRequestDTO {
-    private String email;
-    private String otp;
-}
