--- src/main/java/com/example/Registration_Login/enums/AuthProvider.java
@@ -1,9 +1,0 @@
-package com.example.Registration_Login.enums;
-
-public enum AuthProvider {
-
-    LOCAL, 
-    GOOGLE, 
-    FACEBOOK, 
-    GITHUB
-}
