--- src/main/java/com/example/Registration_Login/enums/Role.java
@@ -1,7 +1,0 @@
-package com.example.Registration_Login.enums;
-
-public enum Role {
-    
-    USER, 
-    ADMIN
-}
