--- src/main/java/com/example/Registration_Login/repository/RefreshTokenRepository.java
@@ -1,13 +1,0 @@
-package com.example.Registration_Login.repository;
-
-import com.example.Registration_Login.model.RefreshToken;
-import com.example.Registration_Login.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
-    Optional<RefreshToken> findByToken(String token);
-
-    void deleteByUser(User user);
-}
