--- src/main/java/com/example/Registration_Login/security/SendGridConfig.java
@@ -1,18 +1,0 @@
-package com.example.Registration_Login.security;
-
-import com.sendgrid.SendGrid;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SendGridConfig {
-
-    @Value("${sendgrid.api-key}")
-    private String apiKey;
-
-    @Bean
-    public SendGrid sendGrid() {
-        return new SendGrid(apiKey);
-    }
-}
