--- src/main/java/com/example/Registration_Login/service/CustomUserDetailsService.java
@@ -1,32 +1,0 @@
-package com.example.Registration_Login.service;
-
-import com.example.Registration_Login.model.User;
-import com.example.Registration_Login.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import java.util.Collections;
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-public class CustomUserDetailsService implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        List<User> users = userRepository.findByEmail(email);
-        if (users.isEmpty()) {
-            throw new UsernameNotFoundException("User not found with email: " + email);
-        }
-        User user = users.get(0);
-        return new org.springframework.security.core.userdetails.User(
-                user.getEmail(),
-                user.getPassword(),
-                Collections.emptyList());
-    }
-}
