--- src/main/java/com/example/Registration_Login/service/RefreshTokenService.java
@@ -1,48 +1,0 @@
-package com.example.Registration_Login.service;
-
-import com.example.Registration_Login.model.RefreshToken;
-import com.example.Registration_Login.model.User;
-import com.example.Registration_Login.repository.RefreshTokenRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-
-import java.time.Instant;
-import java.util.Optional;
-import java.util.UUID;
-
-@Service
-@RequiredArgsConstructor
-public class RefreshTokenService {
-
-    @Value("${jwt.refresh.expiration}")
-    private Long refreshTokenDurationMs;
-
-    private final RefreshTokenRepository refreshTokenRepository;
-
-    public Optional<RefreshToken> findByToken(String token) {
-        return refreshTokenRepository.findByToken(token);
-    }
-
-    public RefreshToken createRefreshToken(User user) {
-        RefreshToken refreshToken = new RefreshToken();
-        refreshToken.setUser(user);
-        refreshToken.setExpiryDate(Instant.now().plusMillis(refreshTokenDurationMs));
-        refreshToken.setToken(UUID.randomUUID().toString());
-
-        refreshToken = refreshTokenRepository.save(refreshToken);
-        return refreshToken;
-    }
-
-    public RefreshToken verifyExpiration(RefreshToken token) {
-        if (token.getExpiryDate().compareTo(Instant.now()) < 0) {
-            refreshTokenRepository.delete(token);
-            throw new RuntimeException("Refresh token was expired. Please make a new signin request");
-        }
-        return token;
-    }
-
-    public void deleteByUser(User user) {
-        refreshTokenRepository.deleteByUser(user);
-    }
-}
