--- src/main/java/com/example/Registration_Login/util/JwtUtil.java
@@ -1,91 +1,0 @@
-package com.example.Registration_Login.util;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.stereotype.Component;
-
-import com.example.Registration_Login.enums.Role;
-
-import java.nio.charset.StandardCharsets;
-import java.security.Key;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-
-@Component
-public class JwtUtil {
-
-    private static final String SECRET_KEY = "f0f3d1c1b4d0287e97c78c8932b0b85b7d3ac9e68e87fd2e8a09f6fbc9ffb6b1";
-
-    private final Key key = Keys.hmacShaKeyFor(SECRET_KEY.getBytes(StandardCharsets.UTF_8));
-
-    @Value("${jwt.access.expiration}")
-    private long jwtAccessExpiration;
-
-    public String extractUsername(String token) {
-        return extractClaim(token, Claims::getSubject);
-    }
-
-    public Date extractExpiration(String token) {
-        return extractClaim(token, Claims::getExpiration);
-    }
-
-    public <T> T extractClaim(String token, Function<Claims, T> claimsResolver) {
-        final Claims claims = extractAllClaims(token);
-        return claimsResolver.apply(claims);
-    }
-
-    public Long extractUserId(String token) {
-        final Claims claims = extractAllClaims(token);
-        Number userId = claims.get("userId", Number.class);
-        return userId != null ? userId.longValue() : null;
-    }
-
-    private Claims extractAllClaims(String token) {
-        return Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token).getBody();
-    }
-
-    private Boolean isTokenExpired(String token) {
-        return extractExpiration(token).before(new Date());
-    }
-
-    public String generateToken(String username, Long userId, Set<Role> roles) {
-        Map<String, Object> claims = new HashMap<>();
-        claims.put("roles", roles.stream()
-                .map(role -> "ROLE_" + role.name())
-                .collect(Collectors.toList()));
-        claims.put("userId", userId);
-        return createToken(claims, username);
-    }
-
-    private String createToken(Map<String, Object> claims, String subject) {
-        return Jwts.builder()
-                .setClaims(claims)
-                .setSubject(subject)
-                .setIssuedAt(new Date(System.currentTimeMillis()))
-                .setExpiration(new Date(System.currentTimeMillis() + jwtAccessExpiration))
-                .signWith(key)
-                .compact();
-    }
-
-    public Boolean validateToken(String token, String username) {
-        final String extractedUsername = extractUsername(token);
-        return (extractedUsername.equals(username) && !isTokenExpired(token));
-    }
-
-    public List<SimpleGrantedAuthority> extractRoles(String token) {
-        Claims claims = extractAllClaims(token);
-        List<String> roles = (List<String>) claims.get("roles");
-        return roles.stream()
-                .map(SimpleGrantedAuthority::new)
-                .collect(Collectors.toList());
-    }
-}
