+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.application.name=Registration_Login
+spring.application.name=ImageIOT
 
 server.port=8080
