--- src/test/java/com/example/Registration_Login/RegistrationLoginApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.Registration_Login;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class RegistrationLoginApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
